add ntt

+++ template/template.hpp
@@ -2,7 +2,7 @@
 
 #include <bits/stdc++.h>
 
-#define rep(i, s, n) for (int i = s; i < (int)(n); i++)
+#define rep(i, s, n) for (int i = (int)(s); i < (int)(n); i++)
 #define rrep(i, s, n) for (int i = (int)(n)-1; i >= (int)(s); i--)
 #define all(v) v.begin(), v.end()
